# expression: implement vectorized evaluation for builtinFromDaysSig