--- a/expression/builtin_time_vec.go
+++ b/expression/builtin_time_vec.go
@@ -481,11 +481,31 @@ func (b *builtinTimeStringTimeDiffSig) vecEvalDuration(input *chunk.Chunk, resul
 }
 
 func (b *builtinFromDaysSig) vectorized() bool {
-	return false
+	return true
 }
 
 func (b *builtinFromDaysSig) vecEvalTime(input *chunk.Chunk, result *chunk.Column) error {
-	return errors.Errorf("not implemented")
+	n := input.NumRows()
+	buf, err := b.bufAllocator.get(types.ETInt, n)
+	if err != nil {
+		return err
+	}
+	defer b.bufAllocator.put(buf)
+	if err := b.args[0].VecEvalInt(b.ctx, input, buf); err != nil {
+		return err
+	}
+
+	result.ResizeTime(n, false)
+	result.MergeNulls(buf)
+	t64s := result.Times()
+	i64s := buf.Int64s()
+	for i := 0; i < n; i++ {
+		if result.IsNull(i) {
+			continue
+		}
+		t64s[i] = types.TimeFromDays(i64s[i])
+	}
+	return nil
 }
 
 func (b *builtinMicroSecondSig) vectorized() bool {